+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix: `BuildContext.ancestors` did not include `Component`s.
 - Fix: classes were not updated when the new Element had no class.
+- Fix: attributes were not updated when the new Element had no attributes.
 
 ## 0.1.1